Refine set of messages with valid TLS history.
Add the magic string "version=Tls" to indicate that a message was
received using TLS. This is used by at least one production
MTA (Hurricate MTA by socketlabs).

Correct a typo to potentially reduce the number of messages that are
reported to have used TLS but were sent in cleartext.